#3 MultiparametricSlide, added attribute self.imsize, check image size >2GB
